Moved form fields to own grouped component folder
