+++ src/components/HelFormFields/index.js
@@ -1,5 +1,9 @@
 import HelTextField from './HelTextField.js'
+import HelAutoComplete from './HelAutoComplete.js'
+import MultiLanguageField from './MultiLanguageField.js'
 
 export {
-    HelTextField
+    HelTextField,
+    HelAutoComplete,
+    MultiLanguageField
 }
